refactor: move `GatewayParams` to `gateway.py`
